--- src/main/java/com/teamharmony/newscommunity/users/dto/SignupDto.java
@@ -1,13 +1,0 @@
-package com.teamharmony.newscommunity.users.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
-public class SignupDto {
-	private String username;
-	private String password;
-}
