+++ src/main/java/com/teamharmony/newscommunity/users/repository/UserRepository.java
@@ -1,4 +1,4 @@
-package com.teamharmony.newscommunity.users.repo;
+package com.teamharmony.newscommunity.users.repository;
 
 import com.teamharmony.newscommunity.users.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
