+++ src/main/java/com/teamharmony/newscommunity/users/repository/UserRoleRepository.java
@@ -1,6 +1,5 @@
-package com.teamharmony.newscommunity.users.repo;
+package com.teamharmony.newscommunity.users.repository;
 
-import com.teamharmony.newscommunity.users.entity.Role;
 import com.teamharmony.newscommunity.users.entity.User;
 import com.teamharmony.newscommunity.users.entity.UserRole;
 import org.springframework.data.jpa.repository.JpaRepository;
