+++ src/main/java/com/teamharmony/newscommunity/users/security/SecurityConfig.java
@@ -2,7 +2,7 @@ package com.teamharmony.newscommunity.users.security;
 
 import com.teamharmony.newscommunity.users.filter.CustomAuthenticationFilter;
 import com.teamharmony.newscommunity.users.filter.CustomAuthorizationFilter;
-import com.teamharmony.newscommunity.users.repo.TokensRepository;
+import com.teamharmony.newscommunity.users.repository.TokensRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
 import org.springframework.context.annotation.Bean;
